test(core-magistrate-transactions): test for handlers moved into handlers folder
