+++ __tests__/unit/core-magistrate-transactions/handlers/business-registration.test.ts
@@ -10,19 +10,19 @@ import { StateStore } from "@arkecosystem/core-state/src/stores/state";
 import { TransactionHandler } from "@arkecosystem/core-transactions/src/handlers";
 import { TransactionHandlerRegistry } from "@arkecosystem/core-transactions/src/handlers/handler-registry";
 import { Wallets } from "@arkecosystem/core-state";
-import { configManager } from "@packages/crypto/src/managers";
+import { configManager } from "@arkecosystem/crypto/src/managers";
 import {
     buildSenderWallet,
     initApp,
-} from "./__support__/app";
-import { setMockTransaction } from "./__mocks__/transaction-repository";
+} from "../__support__/app";
+import { setMockTransaction } from "../__mocks__/transaction-repository";
 // import { setMockBlock } from "../__mocks__/block-repository";
 import { BusinessRegistrationBuilder } from "@arkecosystem/core-magistrate-crypto/src/builders";
 import { IBusinessRegistrationAsset } from "@arkecosystem/core-magistrate-crypto/src/interfaces";
 import { Enums, Transactions as MagistrateTransactions } from "@arkecosystem/core-magistrate-crypto";
 import { Handlers } from "@arkecosystem/core-magistrate-transactions";
 import { MagistrateApplicationEvents } from "@arkecosystem/core-magistrate-transactions/src/events";
-import { setMockBlock } from "./__mocks__/block-repository";
+import { setMockBlock } from "../__mocks__/block-repository";
 import {
     BusinessAlreadyRegisteredError,
 } from "@arkecosystem/core-magistrate-transactions/dist/errors";
