+++ __tests__/unit/core-magistrate-transactions/handlers/general.test.ts
@@ -10,9 +10,9 @@ import { StateStore } from "@arkecosystem/core-state/src/stores/state";
 import { TransactionHandler } from "@arkecosystem/core-transactions/src/handlers";
 import { TransactionHandlerRegistry } from "@arkecosystem/core-transactions/src/handlers/handler-registry";
 import { Wallets } from "@arkecosystem/core-state";
-import { configManager } from "@packages/crypto/src/managers";
-import { buildSenderWallet, initApp } from "./__support__/app";
-import { setMockTransaction } from "./__mocks__/transaction-repository";
+import { configManager } from "@arkecosystem/crypto/src/managers";
+import { buildSenderWallet, initApp } from "../__support__/app";
+import { setMockTransaction } from "../__mocks__/transaction-repository";
 import { BusinessRegistrationBuilder } from "@arkecosystem/core-magistrate-crypto/src/builders";
 import { IBusinessRegistrationAsset } from "@arkecosystem/core-magistrate-crypto/src/interfaces";
 import { Enums } from "@arkecosystem/core-magistrate-crypto";
